# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Ticketz LeadEngine API Contracts
  version: 1.0.0
  description: >-
    Especificação mínima das rotas responsáveis pelo envio de mensagens via
    Ticketz LeadEngine API.
  contact:
    name: Ticketz LeadEngine Team
    email: suporte@ticketz.local
servers:
  - url: https://api.ticketz.local
...
paths:
...
  /api/v1/agreements/{agreementId}:
    parameters:
      - name: agreementId
        in: path
        required: true
        description: Identificador único do acordo.
        schema:
          type: string
    get:
      operationId: getAgreement
      summary: Recupera detalhes de um acordo comercial
      description: Retorna o acordo completo com tabelas, janelas, taxas e metadados vigentes.
      tags: [agreements]
      responses:
        '200':
          description: Dados completos do acordo solicitado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgreementResponse'
        '401':
          description: Requisição não autenticada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operação não autorizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Acordo não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    patch:
      operationId: updateAgreement
      summary: Atualiza parcialmente um acordo comercial
      description: Permite ajustes pontuais no acordo, como mudança de status, vigência ou definição de novas taxas.
      tags: [agreements]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgreementUpdateRequest'
      responses:
        '200':
          description: Acordo atualizado com sucesso.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgreementResponse'
        '400':
          description: Payload inválido para atualização.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '401':
          description: Requisição não autenticada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operação não autorizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Acordo não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflito ao atualizar o acordo (ex. vigência sobreposta).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Falhas de validação no payload enviado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
...
  /api/v1/agreements/providers/{providerId}/sync:
    post:
      operationId: triggerProviderAgreementSync
      summary: Dispara sincronização de acordos para um provedor
      description: Agenda uma sincronização assíncrona dos acordos ativos com o provedor remoto.
      tags: [agreements]
      parameters:
        - name: providerId
          in: path
          required: true
          description: Identificador do provedor de integrações.
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProviderSyncRequest'
      responses:
        '202':
          description: Sincronização agendada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProviderSyncResponse'
        '401':
          description: Requisição não autenticada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Operação não autorizada.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Provedor não encontrado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Já existe uma sincronização em andamento para o provedor informado.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
    RequestMeta:
      type: object
      description: Metadados opcionais enviados pelo consumidor para rastreabilidade.
      properties:
        correlationId:
          type: string
          description: Identificador fornecido pelo cliente para correlacionar requisições.
        source:
          type: string
          description: Nome do sistema cliente que originou a requisição.
      additionalProperties: false
    ResponseMeta:
      type: object
      description: Metainformações retornadas pelo serviço em todas as respostas de sucesso.
      properties:
        requestId:
          type: string
          description: Identificador único da requisição gerado pelo serviço.
        generatedAt:
          type: string
          format: date-time
          description: Timestamp em que a resposta foi gerada (ISO 8601, UTC).
        version:
          type: string
          description: Versão da API que produziu a resposta.
      additionalProperties: false
...
    AgreementStatus:
      type: string
      enum: [draft, active, suspended, terminated, archived]
      description: Situação operacional do acordo.
    AgreementRateType:
      type: string
      enum: [flat, percentage, tiered]
      description: Identifica a forma de cálculo aplicada pela taxa.
    AgreementRateTier:
      type: object
      required: [fromVolume, toVolume, value]
      properties:
        id:
          type: string
        fromVolume:
          type: integer
          minimum: 0
        toVolume:
          type: [integer, 'null']
          description: Volume máximo (inclusive) suportado pela faixa. Null indica ausência de limite superior.
          minimum: 0
        value:
          type: number
          description: Valor aplicado para a faixa informada.
        currency:
          type: string
          description: Código da moeda ISO-4217 aplicável quando a faixa utiliza valor fixo.
      additionalProperties: false
    AgreementRateBase:
      type: object
      required: [type]
      properties:
        type:
          $ref: '#/components/schemas/AgreementRateType'
        description:
          type: string
        value:
          type: number
          description: Valor absoluto aplicado para taxas do tipo flat ou percentual (0-100).
        currency:
          type: string
          description: Código da moeda ISO-4217 quando aplicável.
        percentageDecimals:
          type: integer
          minimum: 0
          maximum: 4
          description: Precisão decimal para taxas percentuais.
        minimumValue:
          type: number
        maximumValue:
          type: number
        tiers:
          type: array
          items:
            $ref: '#/components/schemas/AgreementRateTier'
          description: Coleção de faixas aplicável quando type=tiered.
        windowIds:
          type: array
          items:
            type: string
          description: Identificadores das janelas às quais a taxa se aplica.
      additionalProperties: false
    AgreementRate:
      allOf:
        - $ref: '#/components/schemas/AgreementRateBase'
        - type: object
          required: [id]
          properties:
            id:
              type: string
          additionalProperties: false
    AgreementRateInput:
      allOf:
        - $ref: '#/components/schemas/AgreementRateBase'
        - type: object
          properties:
            id:
              type: string
          additionalProperties: false
    AgreementWindowBase:
      type: object
      required: [label, timezone, daysOfWeek, startTime, endTime]
      properties:
        label:
          type: string
          description: Nome exibido da janela.
        timezone:
          type: string
          description: Timezone IANA aplicado às janelas.
        daysOfWeek:
          type: array
          minItems: 1
          items:
            type: string
            enum: [monday, tuesday, wednesday, thursday, friday, saturday, sunday]
        startTime:
          type: string
          pattern: '^[0-2][0-9]:[0-5][0-9]$'
          description: Horário de início no formato HH:mm.
        endTime:
          type: string
          pattern: '^[0-2][0-9]:[0-5][0-9]$'
          description: Horário de término no formato HH:mm.
        startDate:
          type: string
          format: date
        endDate:
          type: [string, 'null']
          format: date
        attributes:
          type: object
          additionalProperties: true
          description: Metadados adicionais específicos da janela.
      additionalProperties: false
    AgreementWindow:
      allOf:
        - $ref: '#/components/schemas/AgreementWindowBase'
        - type: object
          required: [id]
          properties:
            id:
              type: string
          additionalProperties: false
    AgreementWindowInput:
      allOf:
        - $ref: '#/components/schemas/AgreementWindowBase'
        - type: object
          properties:
            id:
              type: string
          additionalProperties: false
    AgreementTableBase:
      type: object
      required: [name, currency, rates]
      properties:
        name:
          type: string
        description:
          type: string
        externalCode:
          type: string
        currency:
          type: string
          description: Código da moeda ISO-4217 utilizada como base para os valores.
        priority:
          type: integer
          description: Ordem de aplicação da tabela.
        windows:
          type: array
          items:
            $ref: '#/components/schemas/AgreementWindowInput'
        rates:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AgreementRateInput'
        metadata:
          type: object
          additionalProperties: true
      additionalProperties: false
    AgreementTable:
      allOf:
        - $ref: '#/components/schemas/AgreementTableBase'
        - type: object
          required: [id]
          properties:
            id:
              type: string
            windows:
              type: array
              items:
                $ref: '#/components/schemas/AgreementWindow'
            rates:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/AgreementRate'
          additionalProperties: false
    AgreementTableInput:
      allOf:
        - $ref: '#/components/schemas/AgreementTableBase'
        - type: object
          properties:
            id:
              type: string
          additionalProperties: false
    AgreementDetail:
      type: object
      required:
        - id
        - providerId
        - name
        - status
        - version
        - validFrom
        - tables
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        providerId:
          type: string
        externalCode:
          type: [string, 'null']
        name:
          type: string
        description:
          type: [string, 'null']
        status:
          $ref: '#/components/schemas/AgreementStatus'
        version:
          type: integer
          minimum: 1
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: [string, 'null']
          format: date-time
        tags:
          type: array
          items:
            type: string
        tables:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/AgreementTable'
        metadata:
          type: object
          additionalProperties: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      additionalProperties: false
...
    AgreementResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/AgreementDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      additionalProperties: false
...
    AgreementUpdatePayload:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/AgreementStatus'
        validFrom:
          type: string
          format: date-time
        validUntil:
          type: [string, 'null']
          format: date-time
        tags:
          type: array
          items:
            type: string
        tables:
          type: array
          items:
            $ref: '#/components/schemas/AgreementTableInput'
        metadata:
          type: object
          additionalProperties: true
      additionalProperties: false
    AgreementUpdateRequest:
      type: object
      required: [data]
      properties:
        data:
          $ref: '#/components/schemas/AgreementUpdatePayload'
        meta:
          $ref: '#/components/schemas/RequestMeta'
      additionalProperties: false
...
    ProviderSyncRequest:
      type: object
      properties:
        forceFullRefresh:
          type: boolean
          description: Força uma sincronização completa, ignorando caches.
        correlationId:
          type: string
        requestedBy:
          type: string
          description: Identificador do operador que iniciou a sincronização.
      additionalProperties: false
    ProviderSyncStatus:
      type: object
      required: [providerId, jobId, status, triggeredAt]
      properties:
        providerId:
          type: string
        jobId:
          type: string
        status:
          type: string
          enum: [queued, syncing, completed, failed]
        triggeredAt:
          type: string
          format: date-time
        lastSuccessfulSyncAt:
          type: [string, 'null']
          format: date-time
        estimatedCompletionAt:
          type: [string, 'null']
          format: date-time
        warnings:
          type: array
          items:
            type: string
        message:
          type: string
      additionalProperties: false
    ProviderSyncResponse:
      type: object
      required: [data, meta]
      properties:
        data:
          $ref: '#/components/schemas/ProviderSyncStatus'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
      additionalProperties: false
...
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              description: Informações adicionais sobre o erro
          additionalProperties: true
      additionalProperties: false
    ValidationErrorIssue:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        path:
          type: array
          items:
            type: string
      additionalProperties: true
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            error:
              type: object
              required: [code, message, details]
              properties:
                code:
                  type: string
                  enum: [VALIDATION_ERROR]
                message:
                  type: string
                details:
                  type: array
                  items:
                    $ref: '#/components/schemas/ValidationErrorIssue'
              additionalProperties: false